requirements.txt
absl-py==0.4.0
aioredis==1.1.0
alabaster==0.7.10
APScheduler==3.4.0
asgiref==2.3.2
astor==0.7.1
async-timeout==2.0.1
attrs==18.2.0
autobahn==18.9.1
Automat==0.7.0
Babel==2.5.1
backcall==0.1.0
backports.weakref==1.0.post1
bleach==1.5.0
boto3==1.4.7
botocore==1.7.43
certifi==2018.8.24
channel==0.1.0
channels==2.1.3
channels-redis==2.3.0
chardet==3.0.4
click==6.7
cloudpickle==0.5.1
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.12.0
configparser==3.5.0
constantly==15.1.0
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
daphne==2.2.2
decorator==4.1.2
dill==0.2.7.1
Django==1.11.6
django-jet==1.0.6
django-js-asset==1.1.0
django-mptt==0.8.7
django-nested-admin==3.0.21
django-solo==1.1.2
djangorestframework==3.8.2
docopt==0.6.2
docutils==0.14
duckling==1.7.3
en-core-web-md==2.0.0
entrypoints==0.2.3
enum34==1.1.6
fakeredis==0.9.0
fbmessenger==5.2.0
Flask==0.12.2
Flask-Cors==3.0.4
ftfy==4.4.3
funcsigs==1.0.2
future==0.16.0
futures==3.1.1
gast==0.2.0
gevent==1.2.2
graphviz==0.8.1
greenlet==0.4.12
grpcio==1.14.1
h5py==2.7.1
hiredis==0.2.0
html5lib==0.999999999
humanfriendly==4.16.1
hyperlink==18.0.0
idna==2.7
imagesize==0.7.1
incremental==17.5.0
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
JPype1==0.6.2
jsonpickle==0.9.5
jsonschema==2.6.0
jupyter-core==4.4.0
Keras==2.1.0
klein==17.10.0
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
mattermostwrapper==2.1
mistune==0.8.1
mock==2.0.0
msgpack==0.5.6
msgpack-numpy==0.4.1
msgpack-python==0.4.8
murmurhash==0.28.0
mysqlclient==1.3.13
nbconvert==5.3.1
nbformat==4.4.0
nbsphinx==0.2.17
networkx==2.0
numpy==1.13.3
packaging==17.1
pandoc==1.0.2
pandocfilters==1.4.2
parso==0.3.1
pathlib==1.0.1
pbr==3.1.1
pexpect==4.6.0
pickleshare==0.7.4
pkg-resources==0.0.0
plac==0.9.6
ply==3.10
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.4.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pykwalify==1.6.0
pyparsing==2.2.0
PySocks==1.6.8
python-crfsuite==0.9.6
python-dateutil==2.6.1
python-monkey-business==1.0.0
python-telegram-bot==10.1.0
pytz==2018.5
PyYAML==3.12
rasa-core==0.10.4
rasa-nlu==0.13.1
redis==2.10.6
regex==2017.4.5
requests==2.19.1
ruamel.yaml==0.15.37
s3transfer==0.1.11
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
sklearn==0.0
sklearn-crfsuite==0.3.6
slackclient==1.2.1
snowballstemmer==1.2.1
spacy==2.0.12
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
tabulate==0.8.2
tensorboard==1.8.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
testpath==0.3.1
thinc==6.10.3
toolz==0.8.2
tqdm==4.23.3
traitlets==4.3.2
twilio==6.14.0
Twisted==18.7.0
txaio==18.8.1
typing==3.6.4
tzlocal==1.4
ujson==1.35
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.51.0
Werkzeug==0.12.2
wrapt==1.10.11
zope.interface==4.5.0
